# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bet AI Pro Platform",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/leagues": {
      "get": {
        "summary": "List Allowed Leagues",
        "operationId": "list_allowed_leagues_api_leagues_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/fixtures/upcoming": {
      "get": {
        "summary": "List Upcoming Fixtures",
        "operationId": "list_upcoming_fixtures_api_fixtures_upcoming_get",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 14,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fixtures/history/sync-missing": {
      "post": {
        "summary": "Sync Missing Fixture History",
        "description": "Queue a bounded history backfill after the fixture panel is loaded.",
        "operationId": "sync_missing_fixture_history_api_fixtures_history_sync_missing_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Sync Missing Fixture History Api Fixtures History Sync Missing Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/fixtures/{fixture_id}/prefill": {
      "get": {
        "summary": "Fixture Prefill",
        "operationId": "fixture_prefill_api_fixtures__fixture_id__prefill_get",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fixture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fixture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyze": {
      "post": {
        "summary": "Analyze Manual",
        "operationId": "analyze_manual_api_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/predict/tiered": {
      "post": {
        "summary": "Predict With Tiered Model",
        "description": "Run the signed tier bundle for research; never emit an actionable bet.",
        "operationId": "predict_with_tiered_model_api_predict_tiered_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TieredPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Predict With Tiered Model Api Predict Tiered Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/predictions/batch": {
      "post": {
        "summary": "Batch Predict",
        "operationId": "batch_predict_api_predictions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 20,
                "title": "Fixture Ids"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Batch Predict Api Predictions Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/analyze/preview": {
      "post": {
        "summary": "Preview Analysis Inputs",
        "description": "Return every point-in-time model input without persisting a prediction.",
        "operationId": "preview_analysis_inputs_api_analyze_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/analyze/fixture/{fixture_id}": {
      "post": {
        "summary": "Analyze Fixture",
        "operationId": "analyze_fixture_api_analyze_fixture__fixture_id__post",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fixture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fixture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ml/status": {
      "get": {
        "summary": "Get Ml Status",
        "operationId": "get_ml_status_api_ml_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TieredPredictionRequest": {
        "properties": {
          "league_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "League Id"
          },
          "features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Features"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "league_id"
        ],
        "title": "TieredPredictionRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bet_ai_access"
      }
    }
  }
}